requirements.txt
flask==3.0.0
torch
torchvision
Pillow==10.1.0
